chore: Move `eventTarget` util to new file
