+++ web/packages/teleport/src/components/UserMenuNav/UserMenuNav.tsx
@@ -26,6 +26,8 @@ import { getCurrentTheme, getNextTheme } from 'design/ThemeProvider';
 
 import { Theme } from 'gen-proto-ts/teleport/userpreferences/v1/theme_pb';
 
+import { focusOutsideTarget } from 'teleport/lib/util/eventTarget';
+
 import session from 'teleport/services/websession';
 import { useFeatures } from 'teleport/FeaturesContext';
 import { useTeleport } from 'teleport';
@@ -41,7 +43,6 @@ import {
   INCREMENT_TRANSITION_DELAY,
 } from 'teleport/components/Dropdown';
 import { DeviceTrustStatus } from 'teleport/TopBar/DeviceTrustStatus';
-import { focusOutsideTarget } from 'teleport/lib/util';
 
 interface UserMenuNavProps {
   username: string;
