+++ web/packages/teleport/src/lib/util/eventTarget.ts
@@ -1,0 +1,31 @@
+/**
+ * Teleport
+ * Copyright (C) 2024  Gravitational, Inc.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import type { FocusEvent } from 'react';
+
+/**
+ * Checks if a focus event occured outside target element
+ */
+export const focusOutsideTarget = (
+  event: FocusEvent<Element>,
+  targetElement: Element
+) =>
+  // `event.currentTarget` is the element the listener is attached to,
+  // `event.relatedTarget` is the element gaining focus
+  !event.currentTarget.contains(event.relatedTarget as Node) &&
+  !targetElement.contains(event.relatedTarget as Node);
